backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.7",
    "langchain>=0.3.13",
    "python-dotenv>=1.0.1",
    "langgraph>=0.2.60",
    "flask>=3.1.0",
    "langchain-openai>=0.2.14",
    "langchain-community>=0.3.13",
    "pinecone[grpc]>=5.4.2",
    "langchain-pinecone>=0.0.1",
]
